Move all test_suite code under zaplib/web/test_suite
To make it a bit easier to understand what is going on.

Also added a note in the docs on how to test with CEF.
