+++ zaplib/web/test_suite/test_suite.ts
@@ -1,18 +1,18 @@
 // @ts-ignore
 import TestSuiteWorker from "worker-loader?inline=no-fallback!./test_suite_worker";
 
-import { assertNotNull, Rpc } from "./common";
+import { assertNotNull, Rpc } from "../common";
 import { TestSuiteTests } from "./test_suite_worker";
-import { PostMessageTypedArray, ZapArray } from "./types";
+import { PostMessageTypedArray, ZapArray } from "../types";
 import { zapBufferTests } from "./zap_buffer_test";
-import * as zaplib from "./zaplib_runtime";
+import * as zaplib from "../zaplib_runtime";
 import {
   expect,
   expectDeallocationOrUnregister as _expectDeallocationOrUnregister,
   expectThrowAsync,
   setInTest,
 } from "./test_helpers";
-import { inWorker } from "./type_of_runtime";
+import { inWorker } from "../type_of_runtime";
 
 const expectDeallocationOrUnregister = (buffer: ZapArray) =>
   _expectDeallocationOrUnregister(zaplib.callRust, buffer);
@@ -401,7 +401,7 @@ zaplib
                   await zaplib.callRust("panic");
                 },
                 // TODO(Paras): An exact line number here is kind of annoying. Later we can have some sort of partial matcher.
-                "panicked at 'I am panicking!', zaplib/test_suite/src/main.rs:109:17"
+                "panicked at 'I am panicking!', zaplib/web/test_suite/src/main.rs:109:17"
               );
 
               await checkWasmOffline();
