+++ zaplib/web/test_suite/zap_buffer_test.ts
@@ -1,5 +1,5 @@
-import { ZapParamType } from "./types";
-import { ZapBuffer, classesToExtend, containsZapBuffer } from "./zap_buffer";
+import { ZapParamType } from "../types";
+import { ZapBuffer, classesToExtend, containsZapBuffer } from "../zap_buffer";
 import { expect, expectThrow } from "./test_helpers";
 
 declare global {
